修复音频硬件支持导致的变调杂音

+++ src/engin/Sound.h
@@ -17,7 +17,7 @@ class XSound {
     // 播放位置
     int playpos;
     // pcm声音数据
-    std::vector<float> pcm_data;
+    std::vector<uint32_t> pcm_data;
     // 暂停标识
     bool pauseflag;
     // 播放速率
