+++ src/engin/codec/decoder.h
@@ -28,7 +28,7 @@ class XAudioDecoder {
     virtual ~XAudioDecoder();
     // 解码整个音频
     int decode_audio(const std::shared_ptr<AVFormatContext>& format,
-                     int streamIndex, std::vector<float>& pcm_data);
+                     int streamIndex, std::vector<uint32_t>& pcm_data);
 };
 
 #endif  // X_AUDIO_DECODER_H
